// shared/types/UserRole.ts
export type UserRole =
  | "CEO"
  | "COO"
  | "CTO"
  | "CPO"
  | "CMO"
  | "CFO"
  | "CIO"
  | "CSO"
  | "Product Manager"
  | "Designer"
  | "UX Designer"
  | "UI Designer"
  | "UX Researcher"
  | "Frontend Developer"
  | "Backend Developer"
  | "Fullstack Developer"
  | "Mobile Developer"
  | "Data Engineer"
  | "ML Engineer"
  | "QA Engineer"
  | "Marketer"
  | "Social Media Manager"
  | "Content Creator"
  | "SEO Specialist"
  | "Community Manager"
  | "Sales"
  | "Business Development"
  | "Partnerships"
  | "Customer Success"
  | "HR"
  | "Legal"
  | "Operations Manager"
  | "Finance"
  | "Advisor"
  | "Other";